requirements.txt
beautifulsoup4==4.11.1
contractions==0.1.72
elasticsearch==8.3.3
emoji==2.0.0
kafka-python==2.0.2
lxml==4.9.1
numpy==1.23.2
num2words==0.5.12
pymongo==4.2.0
requests==2.28.1
selenium==4.4.3
torch==1.12.1
transformers==4.21.1